refactor: renamed to cache reducer
