+++ src/reducer.js
@@ -8,7 +8,7 @@ import {
   errorsReducer,
   queriesReducer,
   crossSliceReducer,
-  optimisticReducer,
+  cacheReducer,
 } from './reducers';
 
 /**
@@ -30,7 +30,7 @@ const combinedReducers = combineReducers({
   listeners: listenersReducer,
   errors: errorsReducer,
   queries: queriesReducer,
-  optimistic: optimisticReducer,
+  cache: cacheReducer,
   composite: state => state, // mock reducer to retain info created by cross slice reducer
 });
